#1 - auth header added to base get func; rename of functions and related files
