+++ src/services/api/GetDomainStatus.ts
@@ -1,0 +1,11 @@
+import { get } from './base';
+
+export function getDomainStatus(title){
+  const options = {
+    params:
+      {
+        domain: title
+      }
+  }
+  return get('domain_status', options)
+}
